requirements.txt
python-dotenv
pymysql[rsa]
pylint
black
pytest
uuid